Layout changes, next/previous buttons and new input fields
Material Design was included using the Material Design Bootstrap css/js library. A new input field with stars is included, which allows for custom ratings of the anime. Anime Adding itself still is not functional.

Next and Previous buttons were added so that you can go back and forward in the steps to add your anime. The next button does not function unless all the inputs are filled in on that page.